# web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mosic 2.0 API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/cipeodg/{id}": {
            "delete": {
                "tags": [
                    "Cipe OdG"
                ],
                "summary": "Eliminazione OdG di un Cipe",
                "operationId": "idCipeOdg",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id dell'OdG",
                        "required": true,
                        "type": "integer",
                        "items": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operazione avvenuta con successo"
                    },
                    "401": {
                        "description": "Autorizzazione negata"
                    }
                }
            }
        },
...
    },
...
}